% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ByProgram.R
\name{ByProgram}
\alias{ByProgram}
\title{ByProgram}
\usage{
ByProgram(Year, with_summary = TRUE)
}
\arguments{
\item{Year}{is digit numeric number of the year}

\item{with_summary}{defaults to TRUE. Adds a summary (total) row and the bottom of the table.}
}
\description{
Creates an table of wrangled data from the Physics Teacher Education Coalition websites organized by Program(e.g. University of Arizona, Brigham Young University idaho, etc.)
}
