% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ByProgramType.R
\name{ByProgramType}
\alias{ByProgramType}
\title{My function that does this}
\usage{
ByProgramType(Year, with_summary = TRUE)
}
\arguments{
\item{Year}{is digit numeric number of the year}

\item{with_summary}{defaults to TRUE. Adds a summary (total) row and the bottom of the table.}
}
\description{
My function that does this
}
