% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicsEdData.R
\name{PhysicsEdData}
\alias{PhysicsEdData}
\title{PhysicsEdData}
\usage{
PhysicsEdData(Year, ByProgram = FALSE, ByProgramFileName = "pfilename",
  ByProgramType = FALSE, ByProgramTypeFileName = "ptfilename",
  ByState = FALSE, ByStateFileName = "sfilename")
}
\arguments{
\item{Year}{is digit numeric number of the year}

\item{ByProgram}{is boolean to create csv file for ByProgram}

\item{ByProgramFileName}{is character, if ByProgram is TRUE the file will receive this name}

\item{ByProgramType}{is boolean to create csv file for ByProgramType}

\item{ByProgramTypeFileName}{is character, if ByProgramType is TRUE the file will receive this name}

\item{ByState}{is boolean to create csv file for ByProgram}

\item{ByStateFileName}{is character, if ByState is TRUE the file will receive this name}
}
\description{
Creates a CSV file for desired version of wrangled data from the Physics Teacher Education Coalition report card website
}
