% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepData.R
\name{PrepData}
\alias{PrepData}
\title{PrepData}
\usage{
PrepData(Year)
}
\arguments{
\item{Year}{is digit numeric number of the year}

\item{with_summary}{defaults to TRUE. Adds a summary (total) row and the bottom of the table.}
}
\description{
Creates an table of wrangled data from the Physics Teacher Education Coalition website to prepare basic data of counts in each category(Major, Subject, Area) and in each program
}
