% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ByProgramType.R
\name{ByProgramType}
\alias{ByProgramType}
\title{ByProgramType}
\usage{
ByProgramType(Year, with_summary = TRUE)
}
\arguments{
\item{Year}{is digit numeric number of the year}

\item{with_summary}{defaults to TRUE. Adds a summary (total) row and the bottom of the table.}
}
\description{
Creates an table of wrangled data from the Physics Teacher Education Coalition websites organized by Program Type(Traditional, Alternative IHE-based or Alternative not IHE-based)
}
